+++ app/controllers/staff/activity_financials_controller.rb
@@ -13,7 +13,7 @@ class Staff::ActivityFinancialsController < Staff::BaseController
 
     @activity = ActivityPresenter.new(activity)
     @transaction_presenters = @transactions.includes(:parent_activity).map { |transaction| TransactionPresenter.new(transaction) }
-    @budget_presenters = @budgets.includes(:parent_activity).map { |budget| BudgetPresenter.new(budget) }
+    @budget_presenters = @budgets.includes(:parent_activity, :providing_organisation).map { |budget| BudgetPresenter.new(budget) }
     @planned_disbursement_presenters = @planned_disbursements.map { |planned_disbursement| PlannedDisbursementPresenter.new(planned_disbursement) }
     @implementing_organisation_presenters = activity.implementing_organisations.map { |implementing_organisation| ImplementingOrganisationPresenter.new(implementing_organisation) }
     @transfers = activity.source_transfers.map { |transfer| TransferPresenter.new(transfer) }