+++ app/controllers/staff/budgets_controller.rb
@@ -65,7 +65,11 @@ class Staff::BudgetsController < Staff::BaseController
       :value,
       :financial_year,
       :currency,
-      :funding_type
+      :funding_type,
+      :providing_organisation_id,
+      :providing_organisation_name,
+      :providing_organisation_type,
+      :providing_organisation_reference
     )
   end