+++ app/presenters/activity_csv_presenter.rb
@@ -4,7 +4,7 @@ class ActivityCsvPresenter < ActivityPresenter
     to_model.intended_beneficiaries.map { |item| I18n.t("activity.recipient_country.#{item}") }.join("; ")
   end
 
-  def beis_id
+  def beis_identifier
     super.to_s
   end