+++ config/environments/production.rb
@@ -31,7 +31,6 @@ Rails.application.configure do
   config.public_file_server.enabled = ENV["RAILS_SERVE_STATIC_FILES"].present?
 
   # Compress JavaScripts and CSS.
-  config.assets.js_compressor = :uglifier
   # Set a css_compressor so sassc-rails does not overwrite the compressor
   # See https://github.com/DFE-Digital/dfe-teachers-payment-service/commit/74ec587cfbe9aa6d0df01a72e99d70ffe9024748
   config.assets.css_compressor = nil