+++ db/migrate/20210413123839_add_providing_organisation_association_to_budget.rb
@@ -1,0 +1,5 @@
+class AddProvidingOrganisationAssociationToBudget < ActiveRecord::Migration[6.0]
+  def change
+    add_reference :budgets, :providing_organisation, type: :uuid, foreign_key: {to_table: :organisations}
+  end
+end