+++ spec/factories/budget.rb
@@ -20,14 +20,19 @@ FactoryBot.define do
 
     trait :transferred do
       budget_type { Budget::BUDGET_TYPES["transferred"] }
+      association :providing_organisation, factory: :delivery_partner_organisation
     end
 
     trait :external_official_development_assistance do
       budget_type { Budget::BUDGET_TYPES["external_official_development_assistance"] }
+      providing_organisation_name { Faker::Company.name }
+      providing_organisation_type { "Other NGO" }
     end
 
     trait :external_non_official_development_assistance do
       budget_type { Budget::BUDGET_TYPES["external_non_official_development_assistance"] }
+      providing_organisation_name { Faker::Company.name }
+      providing_organisation_type { "Other NGO" }
     end
   end
 end