+++ spec/features/staff/users_can_create_a_budget_spec.rb
@@ -7,6 +7,7 @@ RSpec.describe "Users can create a budget" do
     context "when the activity is a fund" do
       scenario "successfully creates a budget" do
         fund_activity = create(:fund_activity, organisation: user.organisation)
+        create(:programme_activity, parent: fund_activity)
 
         visit activities_path
         click_on(fund_activity.title)
@@ -18,7 +19,7 @@ RSpec.describe "Users can create a budget" do
         expect(page).to have_content(t("action.budget.create.success"))
       end
 
-      scenario "a new budget has it's funding type set to that of it's parent activity's source fund" do
+      scenario "a new budget has its funding type set to that of its parent activity's source fund" do
         activity = create(:programme_activity, :gcrf_funded, organisation: user.organisation)
 
         visit activities_path
@@ -30,6 +31,7 @@ RSpec.describe "Users can create a budget" do
 
       scenario "budget creation is tracked with public_activity" do
         fund_activity = create(:fund_activity, organisation: user.organisation)
+        create(:programme_activity, parent: fund_activity)
 
         PublicActivity.with_tracking do
           visit activities_path
@@ -54,8 +56,6 @@ RSpec.describe "Users can create a budget" do
         programme_activity = create(:programme_activity, parent: fund_activity, organisation: user.organisation)
 
         visit activities_path
-        click_on(programme_activity.parent.title)
-        click_on t("tabs.activity.children")
         click_on(programme_activity.title)
 
         click_on(t("page_content.budgets.button.create"))
@@ -71,8 +71,6 @@ RSpec.describe "Users can create a budget" do
 
         visit activities_path
 
-        click_on(programme_activity.parent.title)
-        click_on t("tabs.activity.children")
         click_on(programme_activity.title)
 
         click_on(t("page_content.budgets.button.create"))
@@ -90,8 +88,6 @@ RSpec.describe "Users can create a budget" do
 
         visit activities_path
 
-        click_on(programme_activity.parent.title)
-        click_on t("tabs.activity.children")
         click_on(programme_activity.title)
 
         click_on(t("page_content.budgets.button.create"))
@@ -108,14 +104,20 @@ RSpec.describe "Users can create a budget" do
 
   context "when the user does NOT belong to BEIS" do
     let(:user) { create(:delivery_partner_user) }
+    let(:fund_activity) { create(:fund_activity) }
+    let(:programme_activity) {
+      create(:programme_activity,
+        parent: fund_activity,
+        extending_organisation: user.organisation)
+    }
+    let!(:project_activity) {
+      create(:project_activity,
+        parent: programme_activity,
+        organisation: user.organisation)
+    }
 
     context "on a programme level activity" do
       scenario "they can view but not create budgets" do
-        fund_activity = create(:fund_activity)
-        programme_activity = create(:programme_activity,
-          parent: fund_activity,
-          extending_organisation: user.organisation)
-
         visit organisation_activity_path(user.organisation, programme_activity)
 
         expect(page).to have_content(t("page_content.activity.budgets"))
@@ -124,18 +126,12 @@ RSpec.describe "Users can create a budget" do
     end
 
     context "on a project level activity" do
-      scenario "successfully creates a budget" do
-        fund_activity = create(:fund_activity)
-        programme_activity = create(:programme_activity,
-          parent: fund_activity,
-          extending_organisation: user.organisation)
-        project_activity = create(:project_activity,
-          parent: programme_activity,
-          organisation: user.organisation)
+      scenario "successfully creates a direct budget by default", js: true do
         _report = create(:report, state: :active, organisation: user.organisation, fund: fund_activity)
 
         visit activities_path
 
+        click_on(programme_activity.title)
         click_on(project_activity.title)
 
         click_on(t("page_content.budgets.button.create"))
@@ -145,23 +141,19 @@ RSpec.describe "Users can create a budget" do
         expect(page).to have_content(t("action.budget.create.success"))
       end
 
-      scenario "successfully creates a transferred budget" do
-        fund_activity = create(:fund_activity)
-        programme_activity = create(:programme_activity,
-          parent: fund_activity,
-          extending_organisation: user.organisation)
-        project_activity = create(:project_activity,
-          parent: programme_activity,
-          organisation: user.organisation)
+      scenario "successfully creates a transferred budget", js: true do
         _report = create(:report, state: :active, organisation: user.organisation, fund: fund_activity)
+        another_org = create(:delivery_partner_organisation)
 
         visit activities_path
 
+        click_on(programme_activity.title)
         click_on(project_activity.title)
 
         click_on(t("page_content.budgets.button.create"))
 
         choose("Transferred")
+        choose(another_org.name)
         select "#{Date.current.year}-#{Date.current.next_year.year}", from: "budget[financial_year]"
         fill_in "budget[value]", with: "1000.00"
         click_button t("default.button.submit")
@@ -169,20 +161,49 @@ RSpec.describe "Users can create a budget" do
         expect(page).to have_content(t("action.budget.create.success"))
       end
 
-      scenario "successfully creates an external budget" do
-        fund_activity = create(:fund_activity)
-        programme_activity = create(:programme_activity,
-          parent: fund_activity,
-          extending_organisation: user.organisation)
-        project_activity = create(:project_activity,
-          parent: programme_activity,
-          organisation: user.organisation)
+      scenario "successfully creates an external budget", js: true do
         _report = create(:report, state: :active, organisation: user.organisation, fund: fund_activity)
 
         visit activities_path
 
+        click_on(programme_activity.title)
         click_on(project_activity.title)
 
+        click_on(t("page_content.budgets.button.create"))
+
+        choose("External Official Development Assistance")
+        fill_in("Providing organisation name", with: "Any org in the world")
+        select("International NGO")
+        select "#{Date.current.year}-#{Date.current.next_year.year}", from: "budget[financial_year]"
+        fill_in "budget[value]", with: "1000.00"
+        click_button t("default.button.submit")
+
+        expect(page).to have_content(t("action.budget.create.success"))
+      end
+
+      scenario "for a transferred budget it shows an error if the user doesn't select a providing organisation", js: true do
+        _report = create(:report, state: :active, organisation: user.organisation, fund: fund_activity)
+
+        visit activities_path
+        click_on(programme_activity.title)
+        click_on(project_activity.title)
+        click_on(t("page_content.budgets.button.create"))
+
+        choose("Transferred")
+        select "#{Date.current.year}-#{Date.current.next_year.year}", from: "budget[financial_year]"
+        fill_in "budget[value]", with: "1000.00"
+        click_button t("default.button.submit")
+
+        expect(page).to have_content("There is a problem")
+        expect(page).to have_content t("activerecord.errors.models.budget.attributes.providing_organisation_id.blank")
+      end
+
+      scenario "for an external budget it shows an error if the user doesn't input a providing organisation name and type", js: true do
+        _report = create(:report, state: :active, organisation: user.organisation, fund: fund_activity)
+
+        visit activities_path
+        click_on(programme_activity.title)
+        click_on(project_activity.title)
         click_on(t("page_content.budgets.button.create"))
 
         choose("External Official Development Assistance")
@@ -190,7 +211,48 @@ RSpec.describe "Users can create a budget" do
         fill_in "budget[value]", with: "1000.00"
         click_button t("default.button.submit")
 
-        expect(page).to have_content(t("action.budget.create.success"))
+        expect(page).to have_content("There is a problem")
+        expect(page).to have_content t("activerecord.errors.models.budget.attributes.providing_organisation_name.blank")
+        expect(page).to have_content t("activerecord.errors.models.budget.attributes.providing_organisation_type.blank")
+      end
+
+      context "without JavaScript" do
+        scenario "for a transferred budget it shows an error if the user doesn't select a providing organisation" do
+          _report = create(:report, state: :active, organisation: user.organisation, fund: fund_activity)
+
+          visit activities_path
+          click_on(project_activity.title)
+          click_on(t("page_content.budgets.button.create"))
+
+          choose("Transferred")
+          fill_in("Providing organisation name", with: "Any org in the world")
+          select("International NGO")
+          select "#{Date.current.year}-#{Date.current.next_year.year}", from: "budget[financial_year]"
+          fill_in "budget[value]", with: "1000.00"
+          click_button t("default.button.submit")
+
+          expect(page).to have_content("There is a problem")
+          expect(page).to have_content t("activerecord.errors.models.budget.attributes.providing_organisation_id.blank")
+        end
+
+        scenario "for an external budget it shows an error if the user doesn't input a providing organisation name and type" do
+          _report = create(:report, state: :active, organisation: user.organisation, fund: fund_activity)
+          another_org = create(:delivery_partner_organisation)
+
+          visit activities_path
+          click_on(project_activity.title)
+          click_on(t("page_content.budgets.button.create"))
+
+          choose("External Official Development Assistance")
+          choose(another_org.name)
+          select "#{Date.current.year}-#{Date.current.next_year.year}", from: "budget[financial_year]"
+          fill_in "budget[value]", with: "1000.00"
+          click_button t("default.button.submit")
+
+          expect(page).to have_content("There is a problem")
+          expect(page).to have_content t("activerecord.errors.models.budget.attributes.providing_organisation_name.blank")
+          expect(page).to have_content t("activerecord.errors.models.budget.attributes.providing_organisation_type.blank")
+        end
       end
     end
   end