+++ spec/features/staff/users_can_create_a_third_party_project_level_activity_spec.rb
@@ -5,7 +5,8 @@ RSpec.feature "Users can create a third-party project" do
 
     context "when viewing a project" do
       scenario "a new third party project cannot be added to the programme when a report does not exist" do
-        project = create(:project_activity, :gcrf_funded, organisation: user.organisation)
+        programme = create(:programme_activity, :gcrf_funded, extending_organisation: user.organisation)
+        project = create(:project_activity, :gcrf_funded, organisation: user.organisation, parent: programme)
 
         visit activities_path
         click_on project.title
@@ -15,7 +16,8 @@ RSpec.feature "Users can create a third-party project" do
       end
 
       scenario "a new third party project can be added to the project" do
-        project = create(:project_activity, :gcrf_funded, organisation: user.organisation, extending_organisation: user.organisation)
+        programme = create(:programme_activity, :gcrf_funded, extending_organisation: user.organisation)
+        project = create(:project_activity, :gcrf_funded, organisation: user.organisation, extending_organisation: user.organisation, parent: programme)
         _report = create(:report, state: :active, organisation: user.organisation, fund: project.associated_fund)
 
         visit activities_path
@@ -37,7 +39,8 @@ RSpec.feature "Users can create a third-party project" do
 
       context "without an editable report" do
         scenario "a new third party project cannot be added" do
-          project = create(:project_activity, :gcrf_funded, organisation: user.organisation, extending_organisation: user.organisation)
+          programme = create(:programme_activity, :gcrf_funded, extending_organisation: user.organisation)
+          project = create(:project_activity, :gcrf_funded, organisation: user.organisation, extending_organisation: user.organisation, parent: programme)
 
           visit activities_path
 
@@ -49,7 +52,8 @@ RSpec.feature "Users can create a third-party project" do
       end
 
       scenario "the activity saves its identifier as read-only `transparency_identifier`" do
-        project = create(:project_activity, :gcrf_funded, organisation: user.organisation, extending_organisation: user.organisation)
+        programme = create(:programme_activity, :gcrf_funded, extending_organisation: user.organisation)
+        project = create(:project_activity, :gcrf_funded, organisation: user.organisation, extending_organisation: user.organisation, parent: programme)
         _report = create(:report, state: :active, organisation: user.organisation, fund: project.associated_fund)
         identifier = "3rd-party-proj"
 
@@ -67,7 +71,8 @@ RSpec.feature "Users can create a third-party project" do
       end
 
       scenario "third party project creation is tracked with public_activity" do
-        project = create(:project_activity, :gcrf_funded, organisation: user.organisation, extending_organisation: user.organisation)
+        programme = create(:programme_activity, :gcrf_funded, extending_organisation: user.organisation)
+        project = create(:project_activity, :gcrf_funded, organisation: user.organisation, extending_organisation: user.organisation, parent: programme)
         _report = create(:report, state: :active, organisation: user.organisation, fund: project.associated_fund)
 
         PublicActivity.with_tracking do