+++ spec/features/user_can_view_static_pages_spec.rb
@@ -9,6 +9,7 @@ RSpec.feature "Users can view the static pages" do
       expect(page).to have_link t("footer.link.cookie_statement"), href: page_path("cookie_statement")
       expect(page).to have_link t("footer.link.accessibility_statement"), href: page_path("accessibility_statement")
       expect(page).to have_link t("footer.link.terms_of_service"), href: page_path("terms_of_service")
+      expect(page).to have_link t("footer.link.service_performance"), href: "https://beisodahelp.zendesk.com/hc/en-gb/sections/1500001330861-Service-Performance"
       expect(page).to have_link t("footer.link.support_site"), href: "https://beisodahelp.zendesk.com/"
     end
   end